+++ example/i2c_ds1307.dart
@@ -28,7 +28,7 @@ bool confirm() {
 /// exclusively at 5V.
 void main() {
   // Select the right I2C bus number /dev/i2c-?
-  // 1 for Raspberry Pi, 0 for NanoPi (Armbian), 2 Banana Pi (Armbian)
+  // 1 for Raspberry Pi, 0 for NanoPi (Armbian), 2 Banana Pi (Armbian), 4 BPI-F3
   var i2c = I2C(1);
   try {
     print("dart_periphery Version: $dartPeripheryVersion");
@@ -45,7 +45,7 @@ void main() {
       print("RTC on board temperature sensor: ${rtc.getTemperature()}");
     }
 
-    print("Set RTC to current sytem time?");
+    print("Set RTC to current system time?");
     if (!confirm()) {
       return;
     }