+++ example/i2c_pcf8591.dart
@@ -10,7 +10,7 @@ import 'dart:io';
 
 // PCF8591 ADC+DAC combo
 //
-// Tested model form Botland
+// Tested model from Botland
 // https://botland.store/raspberry-pi-gpio-extensions/2632-pcf8591-a-d-and-d-a-converter-8-bit-i2c-v2-5904422359164.html
 // with following build-in sensors thermistor,phototransistor, potentiometer
 //
@@ -20,7 +20,7 @@ import 'dart:io';
 //
 void main(List<String> args) {
   // Select the right I2C bus number /dev/i2c-?
-  // 1 for Raspberry Pi, 0 for NanoPi (Armbian), 2 Banana Pi (Armbian)
+  // 1 for Raspberry Pi, 0 for NanoPi (Armbian), 2 Banana Pi (Armbian), 4 BPI-F3
 
   if (args.isEmpty || !(args[0] == "write" || args[0] == "read")) {
     print("Commands: read or write");