+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.9.16
+* Add VL53L0X Time of Fligth sensor
+* Update to Dart 3.7.0
+* Add platform package switch `linux` only   
+* Enable project [WIKI](https://github.com/pezi/dart_periphery/wiki)
+* Test dart_periphery on a real [RISC-V Board](https://github.com/pezi/dart_periphery/wiki/BPI_F3)
+
 ## 0.9.15
 
 * Add DS1307/DS3231 real time clock support